Move toward an event-based way of checking whether a thrown object is harmful
