+++ Resources/Locale/en-US/pacified/pacified.ftl
@@ -3,6 +3,8 @@
 
 # With projectiles:
 pacified-cannot-throw = You can't bring yourself to throw { THE($projectile) }, that could hurt someone!
+# With embedding projectiles:
+pacified-cannot-throw-embed = No way you can throw { THE($projectile) }, that could get lodged inside someone!
 # With liquid-spilling projectiles:
 pacified-cannot-throw-spill = You can't possibly throw { THE($projectile) }, that could spill nasty stuff on someone!
 # With slippery projectiles:
